requirements.txt
streamlit
boto3
pymongo
python-dotenv
requests
pandas
matplotlib
numpy